requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
beautifulsoup4==4.14.3
browser-use==0.11.7
browser-use-sdk==2.0.14
bubus==1.5.6
cdp-use==1.4.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.4
curl_cffi==0.13.0
distro==1.9.0
docstring_parser==0.17.0
frozendict==2.4.7
frozenlist==1.8.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.61.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
inquirerpy==0.3.4
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.2
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.1
oauthlib==3.3.1
ollama==0.6.1
openai==2.16.0
packaging==26.0
pandas==3.0.0
peewee==3.19.0
pfzy==0.3.4
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
pluggy==1.6.0
portalocker==2.10.1
posthog==7.8.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.11.0
pyotp==2.9.0
pyparsing==3.3.2
pypdf==6.6.2
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
tenacity==9.1.2
text-unidecode==1.3
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid7==0.1.0
uv==0.9.28
uvicorn==0.40.0
wcwidth==0.5.3
websockets==15.0.1
yarl==1.22.0
yfinance==1.1.0
